% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplots-package.R
\docType{package}
\name{nplots}
\alias{nplots}
\alias{nplots-package}
\title{nplots: A convenient way of plotting various plot types based on ggplot2}
\description{
To avoid repetitive writing of code that generates plots, use this awesome package to call functions that plot various plots based off of ggplot2. Version 0.1.0 of this package provides the nplots_scatter() and nplots_violin() functions to generate scatter and violin plots from ggplot2 respectively.
}
\note{
This is version 0.1.0 of nplots (November 2018). Future package development will provide an inclusion of other plot types (like box, line, bar, tile e.t.c) and more plot layers and objects to work with.
}
\seealso{
Useful links:
   \url{http://ggplot2.tidyverse.org}
}
\author{
Rebecca Asiimwe
}
\keyword{nplots}
