% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplots_scatter.R
\name{nplots_scatter}
\alias{nplots_scatter}
\title{nplots_scatter
: Generate scatter plots using ggplot2}
\usage{
nplots_scatter(x, x_variable, y_variable, xlab = "", ylab = "",
  title = "", plotly = TRUE, alpha = 0.2, hjust = 0.5)
}
\arguments{
\item{x}{The dataset to be used}

\item{x_variable}{x-axis variable for the scatter plot}

\item{y_variable}{y-axis variable for thescatter plot}

\item{xlab}{argument to specify x axis label}

\item{ylab}{argument to specify y axis label}

\item{title}{argument to specify the plot title to be applied}

\item{plotly}{argument to specify whether the genereated plot should be rendered as a plotly plot}

\item{alpha}{argument to specify geom_poin opacity or transparency magnitude}

\item{hjust}{argument to specify horizontal justification of the plot title}
}
\value{
scatter plot
}
\description{
This function takes your plot arguments like the data, aesthetics, plot labels, desired layers and creates a scatter plot for you using ggplot2. You don't have to worry about writing a bunch of lines of code. This function takes care ofthat for you!
}
\details{
This function takes plot the specified plot arguments and produces a scatter plot using - \code{\link{nplots_scatter}} with the background machinery being ggplot2. Future versions of this function will provide an inclusion of other plot layers and objects to work with.
}
\examples{
\\code {nplots_scatter(gapminder, gapminder$lifeExp, gapminder$gdpPercap, xlab="lifeExp", ylab="gdpPercap", title="lifeExp Vs GdpPercap", plotly=FALSE)}
\\code {nplots_scatter(iris, iris$Sepal.Width, iris$Sepal.Length, xlab="Sepal.Width", ylab="Sepal.Length", title="Sepal Width Vs Length", plotly=FALSE)}
}
